+++ README.md
@@ -1,6 +1,6 @@
-# github-android-app
+# droidhub-android-app
 
-[![Build/Test](https://github.com/vgaidarji/github-android-app/actions/workflows/main.yml/badge.svg?branch=main)](https://github.com/vgaidarji/github-android-app/actions/workflows/main.yml)
+[![Build/Test](https://github.com/vgaidarji/droidhub-android-app/actions/workflows/main.yml/badge.svg?branch=main)](https://github.com/vgaidarji/droidhub-android-app/actions/workflows/main.yml)
 
 # Project configuration
