+++ app/build.gradle.kts
@@ -44,6 +44,7 @@ android {
 
 dependencies {
     implementation(project(AppConfig.Modules.CORE_REPOSITORY))
+    implementation(project(AppConfig.Modules.FEATURE_PROFILE))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
