+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:name=".GitHubAndroidApplication"
+        android:name=".DroidHubAndroidApplication"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.GitHubAndroidApplication"
+        android:theme="@style/Theme.DroidHubAndroidApplication"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
