+++ app/src/main/kotlin/com/vgaidarji/droidhub/DroidHubAndroidApplication.kt
@@ -1,0 +1,8 @@
+package com.vgaidarji.droidhub
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class DroidHubAndroidApplication : Application() {
+}
