--- app/src/main/kotlin/com/vgaidarji/github/app/GitHubAndroidApplication.kt
@@ -1,8 +1,0 @@
-package com.vgaidarji.github.app
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class GitHubAndroidApplication : Application() {
-}
