--- app/src/main/kotlin/com/vgaidarji/github/app/ui/starred/StarredViewModel.kt
@@ -1,13 +1,0 @@
-package com.vgaidarji.github.app.ui.starred
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class StarredViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is starred Fragment"
-    }
-    val text: LiveData<String> = _text
-}
