+++ core/api/src/main/kotlin/com/vgaidarji/droidhub/api/GitHubUsersApi.kt
@@ -1,0 +1,7 @@
+package com.vgaidarji.droidhub.api
+
+import com.vgaidarji.droidhub.model.GitHubUser
+
+interface GitHubUsersApi {
+    suspend fun getGitHubUser(name: String): GitHubUser
+}
