--- core/api/src/main/kotlin/com/vgaidarji/github/api/GitHubUsersApi.kt
@@ -1,7 +1,0 @@
-package com.vgaidarji.github.api
-
-import com.vgaidarji.github.model.GitHubUser
-
-interface GitHubUsersApi {
-    suspend fun getGitHubUser(name: String): GitHubUser
-}
