--- core/repository/src/main/kotlin/com/vgaidarji/github/repository/GitHubUserRepository.kt
@@ -1,16 +1,0 @@
-package com.vgaidarji.github.repository
-
-import com.vgaidarji.github.api.GitHubUsersApi
-import com.vgaidarji.github.model.GitHubUser
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class GitHubUserRepository @Inject constructor(
-    private val gitHubUsersApi: GitHubUsersApi
-) {
-
-    suspend fun getGitHubUser(name: String): GitHubUser {
-        return gitHubUsersApi.getGitHubUser(name)
-    }
-}
