+++ feature/profile/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.vgaidarji.github.profile.ProfileFragment">
+    tools:context="com.vgaidarji.droidhub.profile.ProfileFragment">
 
     <TextView
         android:id="@+id/text_profile"
